Added more tests, moved tests to separate file
Note that these tests are generated using the program itself, and then manually checked whether they seem accurate. This means that the tests are for confirming that the program does not add entries that definitely should not be in there, and not necessarily for getting *all* entries it should get.